docs: split op_server examples and document SqlxOpStore migrate
